+++ backend/database/migrations/2021_03_12_214501_create_options_table.php
@@ -13,19 +13,21 @@ class CreateOptionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('options', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('question_id');
-            $table->string('subcode');
-            $table->string('text');
-            $table->string('value');
-            $table->timestamps();
-            $table->softDeletes();
-            $table->foreign('question_id')
-                ->references('id')
-                ->on('questions')
-                ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('options')) {
+            Schema::create('options', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('question_id');
+                $table->string('subcode');
+                $table->string('text');
+                $table->string('value');
+                $table->timestamps();
+                $table->softDeletes();
+                $table->foreign('question_id')
+                    ->references('id')
+                    ->on('questions')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
